+++ IGenerator.h
@@ -8,5 +8,5 @@ struct LinearSystem;
 class IGenerator
 {
 public:
-	virtual std::shared_ptr<LinearSystem> Generate(int size, int sparseness, std::vector<double> & x) = 0;
+	virtual std::shared_ptr<LinearSystem> Generate(int size, int sparseness, std::vector<float> & x) = 0;
 };
